Начинаю думать над временем
Работает выбор у авто и персонала и апдейт

+++ src/main/java/com/example/springapp/service/AutoPersonnelService.java
@@ -2,6 +2,7 @@ package com.example.springapp.service;
 
 import com.example.springapp.entity.AutoPersonnelEntity;
 import com.example.springapp.repository.AutoPersonnelRepo;
+import com.example.springapp.repository.AutoRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -12,6 +13,9 @@ public class AutoPersonnelService
 {
   @Autowired
   private AutoPersonnelRepo autoPersonnelRepo;
+
+  @Autowired
+  private AutoRepo autoRepo;
 
   public AutoPersonnelEntity addNewAutoPersonnel(AutoPersonnelEntity personnel) throws RuntimeException
   {
@@ -31,7 +35,34 @@ public class AutoPersonnelService
 
   public Boolean deleteById(Integer id)
   {
-    autoPersonnelRepo.deleteById(id);
-    return true;
+    Optional<AutoPersonnelEntity> autoPersonnel = autoPersonnelRepo.findById(id);
+    if (autoPersonnel.isPresent())
+    {
+      autoPersonnelRepo.deleteById(id);
+      return true;
+    }
+    return false;
+  }
+
+  public AutoPersonnelEntity newName(Integer id, String newName)
+  {
+    Optional<AutoPersonnelEntity> personnel = autoPersonnelRepo.findById(id);
+
+    if (!personnel.isPresent())
+    {
+      throw new RuntimeException("Такого водителя нет в базе");
+    }
+
+
+    Optional<AutoPersonnelEntity> personnelEntity = autoPersonnelRepo.findByFirstNameAndLastNameAndPatherName(newName, personnel.get().getLastName(), personnel.get().getPatherName());
+
+    if (personnelEntity.isPresent())
+    {
+      throw new RuntimeException("Водитель с такими ФИО уже записан");
+
+
+    }
+    personnel.get().setFirstName(newName);
+    return autoPersonnelRepo.save(personnel.get());
   }
 }
